+++ great-eye.js
@@ -1,5 +1,5 @@
 module.exports = {
-  extends: ["ash-nazg/sauron"],
+  extends: ["./sauron"],
   settings: {
     preferredTypes: {
       number: {
@@ -10,10 +10,14 @@ module.exports = {
   rules: {
     "capitalized-comments": ["warn"],
     "complexity": ["warn"],
+    "max-statements": ["warn"],
     "max-statements-per-line": ["warn"],
     "no-magic-numbers": ["warn"],
     "no-plusplus": ["warn"],
     "no-warning-comments": ["warn"],
+
+    "import/no-unused-modules": ["warn", {missingExports: true, unusedExports: true}],
+
     "jsdoc/check-alignment": ["warn"],
     "jsdoc/check-indentation": ["warn"],
     "jsdoc/require-description": ["warn"],
@@ -21,6 +25,9 @@ module.exports = {
     "jsdoc/require-returns-description": ["warn"],
     "jsdoc/require-example": ["warn"],
     "jsdoc/require-jsdoc": ["error"],
+
+    "sonarjs/cognitive-complexity": "warn",
+
     "unicorn/prevent-abbreviations": "error"
   }
 };