+++ index.js
@@ -1,13 +1,58 @@
 module.exports = {
-  "extends": ["eslint:recommended", "plugin:unicorn/recommended", "standard", "plugin:eslint-comments/recommended", "plugin:compat/recommended"],
-  "plugins": ["promise", "import", "unicorn", "no-use-extend-native", "jsdoc", "markdown", "@mysticatea"],
+  "extends": [
+    // This plugin's rules are overridden by following configs (except for
+    //   its own rules), so including first; we also don't want its disabling
+    //   of some, e.g., recommended rules
+    // Actually, we're disabling instead for now (and manually enabling in
+    //   `index.js` and adding to `plugins`), as has disabling and adds rules
+    //   we don't want; filed the following to get an "all" config instead
+    //   showing only rules specific to the plugin:
+    //     https://github.com/mysticatea/eslint-plugin/issues/21
+    // "plugin:@mysticatea/es2019",
+
+    "eslint:recommended",
+    // These may override "standard" which includes their rules,
+    //   so including first
+    "plugin:promise/recommended",
+    "plugin:import/recommended",
+    // This may override eslint:recommended and the promise/import rules,
+    //  but it doesn't turn any off
+    "standard",
+    // These only add their own rules and are not used by the others
+    "plugin:unicorn/recommended",
+    "plugin:eslint-comments/recommended",
+    "plugin:compat/recommended",
+    "plugin:array-func/all",
+    "plugin:sonarjs/recommended",
+    "plugin:jsdoc/recommended"
+  ],
+  "plugins": [
+    // These have no rules
+    "markdown",
+    "html",
+    // See above on why adding here instead of extending a config
+    "@mysticatea",
+    // Filed the following to avoid need for adding `plugins` for this as well
+    //  as adding the rule manually as we do below (will need to reference in
+    //  `package.json` if do so can show in `implicitly-included`):
+    // https://github.com/dustinspecker/eslint-plugin-no-use-extend-native/issues/125
+    "no-use-extend-native",
+    // Filed the following to avoid need for adding `plugins` for this as
+    //  well as `extends`: https://github.com/SonarSource/eslint-plugin-sonarjs/issues/117
+    "sonarjs"
+  ],
   "settings": {
     "jsdoc": {
       "matchingFileName": "dummy.md",
-      "rejectExampleCodeRegex": "^`"
+      "rejectExampleCodeRegex": "^`",
+      "tagNamePreference": {
+        "augments": "extends"
+      }
     }
   },
   "rules": {
+    // RULES SHOULD ONLY BE DIFFERENT FROM INHERITED (ENABLING, DISABLING, CHANGING)
+
     "array-bracket-newline": ["error", "consistent"],
     "array-bracket-spacing": ["error"],
     "array-callback-return": ["error"],
@@ -26,7 +71,6 @@ module.exports = {
     "linebreak-style": ["error"],
     "max-nested-callbacks": ["error"],
     "multiline-ternary": ["error", "always-multiline"],
-    "no-async-promise-executor": ["error"],
     "no-await-in-loop": ["error"],
     "no-bitwise": ["error"],
     "no-buffer-constructor": ["error"],
@@ -41,10 +85,8 @@ module.exports = {
     "no-implicit-globals": ["error"],
     "no-lonely-if": ["error"],
     "no-loop-func": ["error"],
-    "no-misleading-character-class": ["error"],
     "no-mixed-requires": ["error", {"grouping": true, "allowCall": true}],
     "no-process-env": ["error"],
-    "no-prototype-builtins": ["error"],
     "no-restricted-globals": ["error", {
       "name": "event",
       "message": "Use local event parameter instead (preferably as \"e\" or \"ev\")."
@@ -70,7 +112,6 @@ module.exports = {
     "prefer-spread": ["error"],
     "quote-props": ["error", "as-needed"],
     "radix": ["error", "as-needed"],
-    "require-atomic-updates": ["error"],
     "require-await": ["error"],
     "semi-style": ["error"],
     "strict": ["error"],
@@ -84,18 +125,18 @@ module.exports = {
     "object-curly-spacing": ["error", "never"],
     "no-restricted-syntax": ["error", "[operator=instanceof]"],
 
-    "promise/catch-or-return": "error",
-    "promise/no-return-wrap": "error",
-    "promise/always-return": "error",
-
     "promise/no-new-statics": "error",
     "promise/prefer-await-to-then": "error",
 
+    "promise/param-names": "off",
+    "promise/no-native": "off",
+    "promise/no-nesting": "off",
+    "promise/no-promise-in-callback": "off",
+    "promise/no-callback-in-promise": "off",
+    "promise/avoid-new": "off",
+    "promise/no-return-in-finally": "off",
+    "promise/valid-params": "off",
 
-    "import/no-unresolved": "error",
-    "import/named": "error",
-    "import/default": "error",
-    "import/namespace": "error",
     "import/no-absolute-path": "error",
     "import/no-dynamic-require": "error",
     "import/no-self-import": "error",
@@ -117,17 +158,22 @@ module.exports = {
     "import/newline-after-import": "error",
     "import/no-anonymous-default-export": "error",
 
-    "jsdoc/match-description": ["error"],
+    "jsdoc/check-alignment": ["off"],
+    "jsdoc/newline-after-description": ["off"],
+    "jsdoc/no-undefined-types": ["off"],
+    "jsdoc/require-description": ["off"],
+    "jsdoc/require-param-description": ["off"],
+    "jsdoc/require-returns-description": ["off"],
+    "jsdoc/require-example": ["off"],
+    "jsdoc/require-jsdoc": ["off"],
+    "jsdoc/require-returns": ["off"],
+
     "jsdoc/check-examples": ["error"],
     "jsdoc/check-param-names": ["error"],
     "jsdoc/check-syntax": ["error"],
     "jsdoc/check-tag-names": ["error"],
     "jsdoc/check-types": ["error"],
-    "jsdoc/require-param": ["warn"],
-    "jsdoc/require-param-name": ["warn"],
-    "jsdoc/require-param-type": ["warn"],
-    "jsdoc/require-returns-check": ["error"],
-    "jsdoc/require-returns-type": ["warn"],
+    "jsdoc/match-description": ["error"],
     "jsdoc/valid-types": ["error"],
 
     "unicorn/catch-error-name": "off",
@@ -140,15 +186,44 @@ module.exports = {
 
     "unicorn/custom-error-definition": "error",
     "unicorn/no-unsafe-regex": "error",
-    "unicorn/prefer-event-key": "error",
-    "unicorn/prefer-flat-map": "error",
 
     "@mysticatea/block-scoped-var": "error",
+
+    "@mysticatea/no-use-ignored-vars": "off",
+    "@mysticatea/prettier": "off",
+    "@mysticatea/prefer-for-of": "off",
+
+    // These should be inherited by `es2019` config, but we don't want to
+    //   extend as it has other items we don't want (see comment at top)
     "@mysticatea/no-literal-call": "error",
     "@mysticatea/no-this-in-static": "error",
-    "@mysticatea/no-use-ignored-vars": "off",
     "@mysticatea/no-useless-rest-spread": "error",
+
+    // We're directly using `eslint-comments` and without this, these will
+    //  mistakenly show up in our @mysticatea/eslint-plugin inherited list
+    "@mysticatea/eslint-comments/disable-enable-pair": "off",
+    "@mysticatea/eslint-comments/no-aggregating-enable": "off",
+    "@mysticatea/eslint-comments/no-duplicate-disable": "off",
+    "@mysticatea/eslint-comments/no-unlimited-disable": "off",
+    "@mysticatea/eslint-comments/no-unused-disable": "off",
+    "@mysticatea/eslint-comments/no-unused-enable": "off",
+    "@mysticatea/eslint-comments/no-use": "off",
+
     "eslint-comments/no-unused-disable": "error",
-    "eslint-comments/disable-enable-pair": ["error", {"allowWholeFile": true}]
+    "eslint-comments/disable-enable-pair": ["error", {"allowWholeFile": true}],
+
+    "array-func/prefer-array-from": "off",
+
+    "sonarjs/max-switch-cases": "off",
+    "sonarjs/no-collapsible-if": "off",
+    "sonarjs/no-duplicate-string": "off",
+    "sonarjs/no-identical-functions": "off",
+    "sonarjs/no-small-switch": "off",
+    "sonarjs/prefer-immediate-return": "off",
+    "sonarjs/cognitive-complexity": "off",
+
+    // Should not be needed here; see comment in `plugins`
+    //   re: "no-use-extend-native" above
+    "no-use-extend-native/no-use-extend-native": "error"
   }
 };