+++ inherited-rules/implicitly-included/array-func-all.js
@@ -1,0 +1,9 @@
+module.exports = {
+  rules: {
+    "array-func/from-map": "error",
+    "array-func/no-unnecessary-this-arg": "error",
+    "array-func/avoid-reverse": "error",
+    "array-func/prefer-flat-map": "error",
+    "array-func/prefer-flat": "error"
+  }
+};