+++ inherited-rules/implicitly-included/eslint-comments-recommended.js
@@ -1,0 +1,8 @@
+module.exports = {
+  rules: {
+    "eslint-comments/no-aggregating-enable": "error",
+    "eslint-comments/no-duplicate-disable": "error",
+    "eslint-comments/no-unlimited-disable": "error",
+    "eslint-comments/no-unused-enable": "error"
+  }
+};