+++ inherited-rules/implicitly-included/promise-recommended.js
@@ -1,0 +1,7 @@
+module.exports = {
+  rules: {
+    "promise/always-return": "error",
+    "promise/no-return-wrap": "error",
+    "promise/catch-or-return": "error"
+  }
+};