+++ sauron-node.js
@@ -1,5 +1,5 @@
 module.exports = {
-  extends: ["ash-nazg/node", "ash-nazg/sauron"],
+  extends: ["./node", "./sauron"],
   rules: {
     "node/prefer-promises/dns": ["error"],
     "node/prefer-promises/fs": ["error"]